+++ crates/webcodex-validation/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "webcodex-validation"
+version.workspace = true
+license.workspace = true
+edition.workspace = true
+
+[dependencies]
+webcodex-core = { path = "../webcodex-core" }
+webcodex-tool-runtime-contracts = { path = "../webcodex-tool-runtime-contracts" }
+webcodex-workflow-session = { path = "../webcodex-workflow-session" }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+sha2 = "0.10"
+toml = "0.8"
+
