+++ crates/webcodex-validation/src/adapters/go.rs
