+++ crates/webcodex-validation/src/adapters/rust.rs
